+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Update website to dark mode. @SkylerGodfrey
 - Created canvas bounds so player can not leave canvas when moving MyBox. @SkylerGodfrey
 - Created level transition feature, including proper scaling of objects. @SkylerGodfrey
+- Created level 1 objects and finished level 1 design. @SkylerGodfrey
+- Created collision system for objects with 'wallSolid' class. @SkylerGodfrey
  
 ### Changed